Extract Models into Models folder
