+++ MovieQuiz/Models/QuizResultViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  QuizResultViewModel.swift
+//  MovieQuiz
+//
+//  Created by Andrei Chenchik on 11/8/22.
+//
+
+import Foundation
+
+struct QuizResultViewModel {
+    let title: String
+    let text: String
+    let buttonText: String
+}
